+++ pipelex/system/registries/func_registry_utils.py
@@ -87,7 +87,7 @@ class FuncRegistryUtils:
             pattern="*.py",
             is_recursive=is_recursive,
             excluded_dirs=list(get_config().pipelex.scan_config.excluded_dirs),
-            exception_dirs=[folder_path],
+            force_include_dirs=[folder_path],
         )
 
         for python_file in python_files: